Store log lines in memory during tests
When running the go test tool to generate a coverage profile, all log
lines are written to stderr, because of the application's default
logger.

This commit introduces a new type, LineWriter, which buffers lines
separated by `\n` and appends them to a slice of strings.  This writer
is used for the application's logging.

Using LineWriter, logging output is generated when generating a coverage
profile.

Lines are stored in memory so that they are accessible in tests, should
need be.